Use wrapping_sub in bandwidth calculation (#3019)
It happened. I received more than 2^32 GiB while smoldot was running,
and this triggered a panic in the memory printer because the total bytes
transferred overflowed.
The fix is simple: just use `wrapping_sub`.

This panic only happens if debug_assertions are enabled, which is the
case locally but not in the published version, so this panic isn't worth
putting in the CHANGELOG.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>