Modify Merkle proofs to be a single Vec (#3021)
This PR modifies the way Merkle proofs are passed around in the source
code.

Before this PR, the networking code turns the networking response into a
`Vec<Vec<u8>>`, which is then passed around smoldot. The inner `Vec<u8>`
are still encoded, which means that the networking code decodes the
proof a bit but not completely, which is kind of weird.

After this PR, the networking code returns the completely undecoded
proof, and the decoding is now entirely done in the `proof_decode`
module.

This should make everything easier to understand.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>