Clarify meaning of is_unreachable (#3023)
Close https://github.com/paritytech/smoldot/issues/1531
cc https://github.com/paritytech/smoldot/issues/2746

This PR renames `is_unreachable` to `is_bad_address`.
It is generally not possible to know whether an address is definitely
unreachable or not. The only situation where we can affirm that an
address is definitely unreachable is if our implementation doesn't
support a certain protocol that it uses.

This matches what the browser node already does anyway.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>