Implement encoding Merkle proofs (#3033)
cc https://github.com/paritytech/smoldot/issues/2973

I went for an API where the user provides already-existing node values
in order to encode the proof. I went for this, as the main point of
building a proof is to extract information from the trie, and when doing
that you normally already know node values.

I also think that this API is very easy to understand compared to other
APIs that I could have gone for.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>