Move Merkle proof verification to warp syncing module (#3054)
Before this PR, the warp syncing code asks "hey I'd like to know the
runtime", and the upper layers download a Merkle proof, verify it, and
give the storage values to the warp syncing code.

This PR modifies this. The warp syncing code now asks for a Merkle proof
of arbitrary keys, and the upper layers download that proof and give to
the warp syncing code. It is the warp syncing code that verifies said
proof.

The objective of this PR is simplification of the API surface of the
syncing code.
It will also make it easier to implement
https://github.com/paritytech/smoldot/issues/1769, as it is now clear
who is responsible for first downloading only the hash.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>